requirements.txt
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
faster-whisper>=1.0.0
edge-tts>=6.1.0
httpx>=0.24.0
pydub>=0.25.0
python-dotenv>=1.0.0
